requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2
httpx==0.25.1
python-multipart==0.0.6
websockets==12.0
slowapi==0.1.8 